feat: 코어데이터 context 관리 객체 분리
